ragbot/requirements.txt
# requirements.txt (snippet)
chromadb==0.5.5
sentence-transformers==3.0.1
groq
pyyaml
pydantic>=2.6.0,<3
tqdm
tabulate
pytest


openfga-sdk>=0.4.0